fyp-backend-main/fyp-backend-main/integration-chunker/requirements.txt
redis
boto3
pyyaml
langchain_google_genai
langchain_pinecone
langchain
dotenv